Project/project_da/spotify/requirements.txt
spotipy>=2.22.0
pandas
numpy
scikit-learn
matplotlib
python-dotenv
tqdm
streamlit